+++ src/setup-docker/templates/fabric-docker/commands-generated.sh
@@ -214,9 +214,9 @@ networkDown() {
   <% chaincodes.forEach((chaincode) => { -%>
     <% chaincode.channel.orgs.forEach((org) => { -%>
       <% org.peers.forEach((peer) => { -%>
-        <% const chaincodeContainerName=`dev-${peer.address}-${chaincode.name}-${chaincode.version}` -%>
-        docker rm -f $(docker ps -a | grep <%= chaincodeContainerName %>-* | awk '{print $1}') || echo "docker rm failed, Check if all fabric dockers properly was deleted"
-        docker rmi $(docker images <%= chaincodeContainerName %>-* -q) || echo "docker rm failed, Check if all fabric dockers properly was deleted"
+        <% const chaincodeContainerName=`dev-${peer.address}-${chaincode.name}` -%>
+        for container in $(docker ps -a | grep "<%= chaincodeContainerName %>*" | awk '{print $1}'); do echo "Removing container $container..."; docker rm -f "$container" || echo "docker rm of $container failed. Check if all fabric dockers properly was deleted"; done
+        for image in $(docker images "<%= chaincodeContainerName %>*" -q); do echo "Removing image $image..."; docker rmi "$image" || echo "docker rmi of $image failed. Check if all fabric dockers properly was deleted"; done
       <% }) -%>
     <% }) -%>
   <% }) -%>