+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Fixes
 * Support Apple M1 / arm64 architecture
 * Various fixes in channel scripts
+* Remove remaining docker containers and images after prune
 
 ### Chore & Maintenance
 * Add `fabricNodeenvVersion` global configuration